changed user input and fixed some bugs
The developer, manager, and tester pages list each bug in a table with options provided by buttons. The form gives a random ID to the user and has the option to show the client the status of their bug (as well as the info they input, so they can be sure it's the correct bug). The table page lists all bugs overall, and has the option to add/delete tables, which is useful for clearing everything from a table. If the correct tables don't exist, nothing else will work though.